lesson05 TernaryOperator commit lesson 05
