--- src/com/gmail/gloriyaarmen/lessonsJava/lesson05/RelationalDemo.java
@@ -1,15 +1,0 @@
-package com.gmail.gloriyaarmen.lessonsJava.lesson05;
-
-public class RelationalDemo {
-    public static void main(String[] args) {
-
-        int a = 10;
-        int b = -123;
-
-        boolean result = a > b;
-        System.out.println("a > b is " + result);
-        result = a < b;
-        System.out.println("a < b is " + result);
-    }
-
-}
