+++ src/com/gmail/gloriyaarmen/lessonsJava/lesson05/TernaryOperator.java
@@ -1,0 +1,7 @@
+package com.gmail.gloriyaarmen.lessonsJava.lesson05;
+
+public class TernaryOperator {
+    public static void main(String[] args) {
+
+    }
+}
